Extract models into its own module
